move IdDict, IdSet to their own files, rename table.c to iddict.c
